requirements.txt
# ChronoTranscriber Requirements
# Generated: December 2025
# Python 3.13+

# ============================================================================
# Core LLM Integration - LangChain Multi-Provider Support
# ============================================================================
langchain==1.2.4
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-anthropic==1.3.1
langchain-google-genai==4.2.0

# LangChain Dependencies
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
tiktoken==0.12.0

# OpenAI SDK (also used for batch processing)
openai==2.15.0

# Anthropic SDK
anthropic==0.76.0

# Google AI Dependencies
# Google Gen AI SDK (required for Gemini Batch API backends)
google-genai==1.59.0
google-ai-generativelanguage==0.10.0
google-api-core==2.29.0
google-auth==2.47.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
proto-plus==1.27.0
protobuf==6.33.4

# ============================================================================
# Image Processing
# ============================================================================
opencv-python==4.12.0.88
pillow==12.1.0
scikit-image==0.26.0
deskew==1.5.3
imageio==2.37.2

# ============================================================================
# Document Processing
# ============================================================================
PyMuPDF==1.26.7
EbookLib==0.20
mobi==0.4.1
lxml==6.0.2
pytesseract==0.3.13
standard-imghdr==3.13.0

# ============================================================================
# Data Validation & Serialization
# ============================================================================
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3

# ============================================================================
# Async & HTTP
# ============================================================================
aiofiles==25.1.0
aiohttp==3.13.3
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
httpx==0.28.1
httpcore==1.0.9
websockets==15.0.1

# ============================================================================
# Utilities
# ============================================================================
tenacity==9.1.2
colorama==0.4.6
tqdm==4.67.1
loguru==0.7.3

# ============================================================================
# Scientific Computing (dependencies of image processing)
# ============================================================================
numpy==2.2.6
scipy==1.17.0
networkx==3.6.1
tifffile==2026.1.14

# ============================================================================
# Common Dependencies
# ============================================================================
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
docstring_parser==0.17.0
filetype==1.2.0
frozenlist==1.8.0
h11==0.16.0
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
lazy_loader==0.4
multidict==6.7.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
win32_setctime==1.2.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0

# ============================================================================
# Development / Testing
# ============================================================================
# Moved to requirements-dev.txt
